requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
transformers
torch
faiss-cpu
sentence-transformers
python-docx
pypdf
docx2txt
